+++ forge/CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to the format [MCVERSION-MAJORMOD.MAJORAPI.MINOR.PATCH](https://mcforge.readthedocs.io/en/1.15.x/conventions/versioning/).
 
+## [1.17.1-0.29] - 2021.11.12
+- Improved recipe output equality logic
+- Recipe outputs will now display the stack size
+
 ## [1.17.1-0.28] - 2021.11.06
 ### Changed
 - Updated to Minecraft 1.17.1+